requirements.txt
tensorflow>=2.10.0
matplotlib>=3.5.0
numpy>=1.21.0
pandas>=1.3.0
seaborn>=0.11.0
scipy>=1.7.0
Pillow>=9.0.0
scikit-learn>=1.0.0
flask>=2.0.0
gunicorn>=20.1.0
werkzeug>=2.0.0
requests>=2.27.0
google-cloud-storage>=2.0.0